# Test the performance of permutation instances with different parameters