--- a/library/src/permutation/permutation_solution.hpp
+++ b/library/src/permutation/permutation_solution.hpp
@@ -2,7 +2,7 @@
  *
  * MIT License
  *
- * Copyright (C) 2023-2024 Advanced Micro Devices, Inc. All rights reserved.
+ * Copyright (C) 2023-2025 Advanced Micro Devices, Inc. All rights reserved.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -105,6 +105,9 @@ namespace hiptensor
         // Problem dimension
         ck::index_t problemDim() const;
 
+        // Problem size
+        ck::index_t problemSize() const;
+
         // Byte count
         ck::index_t problemBytes() const;
 
@@ -120,6 +123,7 @@ namespace hiptensor
     protected:
         // Derived runtime arguments
         ck::index_t mDim;
+        ck::index_t mSize;
         ck::index_t mBytes;
         bool        mValid;
         uint32_t    mThreadDim;