… CMakeLists.txt
… tuning/CMakeLists.txt
… tuning/permutation/CMakeLists.txt
… tuning/permutation/permutation_instance.hpp
… tuning/permutation/permutation_instance_F16_3_256_128_128_16_16.cpp
… tuning/permutation/permutation_instance_F16_3_256_128_128_32_32.cpp
… tuning/permutation/permutation_instance_F16_3_256_128_128_8_8.cpp
… tuning/permutation/permutation_instance_F16_3_256_64_64_16_16.cpp
… tuning/permutation/permutation_instance_F16_3_256_64_64_4_4.cpp
… tuning/permutation/permutation_instance_F16_3_miscellaneous.cpp
… tuning/permutation/permutation_instance_F16_4_256_128_128_16_16.cpp
… tuning/permutation/permutation_instance_F16_4_256_128_128_32_32.cpp
… tuning/permutation/permutation_instance_F16_4_256_128_128_8_8.cpp
… tuning/permutation/permutation_instance_F16_4_256_64_64_16_16.cpp
… tuning/permutation/permutation_instance_F16_4_256_64_64_4_4.cpp
… tuning/permutation/permutation_instance_F16_4_miscellaneous.cpp
… tuning/permutation/permutation_instance_F16_5_256_128_128_16_16.cpp
… tuning/permutation/permutation_instance_F16_5_256_128_128_32_32.cpp
… tuning/permutation/permutation_instance_F16_5_256_128_128_8_8.cpp
… tuning/permutation/permutation_instance_F16_5_256_64_64_16_16.cpp
… tuning/permutation/permutation_instance_F16_5_256_64_64_4_4.cpp
… tuning/permutation/permutation_instance_F16_5_miscellaneous.cpp
… tuning/permutation/permutation_instance_F16_6_256_128_128_16_16.cpp
… tuning/permutation/permutation_instance_F16_6_256_128_128_32_32.cpp
… tuning/permutation/permutation_instance_F16_6_256_128_128_8_8.cpp
… tuning/permutation/permutation_instance_F16_6_256_64_64_16_16.cpp
… tuning/permutation/permutation_instance_F16_6_256_64_64_4_4.cpp
… tuning/permutation/permutation_instance_F16_6_miscellaneous.cpp
… tuning/permutation/permutation_instance_F32_2_256_128_128_16_16.cpp
… tuning/permutation/permutation_instance_F32_2_256_128_128_32_32.cpp